backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
sqlalchemy
python-dotenv